+++ src/main/java/pe/edu/upc/backendgrupo1/repositories/IRecomendacionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.backendgrupo1.entities.Recomendacion;
 
 @Repository
-public interface IRecomendacionRepository extends JpaRepository<Recomendacion, Long> {
+public interface IRecomendacionRepository extends JpaRepository<Recomendacion, Integer> {
 }
